File reorganization and renaming for the submission flow effort.
Files checked in: DataRepo/templates/DataRepo/validate_submission.html DataRepo/templates/upload.html DataRepo/tests/views/upload/test_validation.py DataRepo/urls.py DataRepo/utils/exceptions.py DataRepo/views/upload/__init__.py DataRepo/views/upload/submission.py DataRepo/views/upload/validation.py TraceBase/settings.py DataRepo/templates/DataRepo/submission.html DataRepo/views/upload/upload.py
