+++ DataRepo/templates/upload.html
@@ -344,11 +344,11 @@
         <br>
     </small>
     <ol>
-        <h5><li><a href="{% url 'validate' %}">Download a Submission Template</a> &nbsp;<a data-bs-toggle="collapse" data-bs-target="#build-submission"><img id="build-submission-button" src="{% static 'images/plus.png' %}" alt="more" width="18"></a></li></h5>
+        <h5><li><a href="{% url 'submission' %}">Download a Submission Template</a> &nbsp;<a data-bs-toggle="collapse" data-bs-target="#build-submission"><img id="build-submission-button" src="{% static 'images/plus.png' %}" alt="more" width="18"></a></li></h5>
         <div id="build-submission" class="major-li collapse">
             <ol>
                 <li>
-                    Go to the <a href="{% url 'validate' %}">Build a TraceBase Submission</a>
+                    Go to the <a href="{% url 'submission' %}">Build a TraceBase Submission</a>
                     page and download an auto-filled template based on your
                     <span class="tooltip-hint" text-tooltip="E.g. AccuCor excel file">peak annotation files</span>.
                     <ul>
@@ -623,11 +623,11 @@
                 </li>
             </ol>
         </div>
-        <h5><li><a href="{% url 'validate' %}">Check for errors</a> &nbsp;<a data-bs-toggle="collapse" data-bs-target="#validate-submission"><img id="validate-submission-button" src="{% static 'images/plus.png' %}" alt="more" width="18"></a></li></h5>
+        <h5><li><a href="{% url 'submission' %}">Check for errors</a> &nbsp;<a data-bs-toggle="collapse" data-bs-target="#validate-submission"><img id="validate-submission-button" src="{% static 'images/plus.png' %}" alt="more" width="18"></a></li></h5>
         <div id="validate-submission" class="major-li collapse">
             <ol>
                 <li>
-                    Go to the <a href="{% url 'validate' %}">Build a TraceBase Submission</a> and submit just your <span class="tooltip-hint" text-tooltip="Animal/Sample excel file">study doc</span> for validation.<br>
+                    Go to the <a href="{% url 'submission' %}">Build a TraceBase Submission</a> and submit just your <span class="tooltip-hint" text-tooltip="Animal/Sample excel file">study doc</span> for validation.<br>
                     <em>While this step is optional, resolving any simple errors on your own can greatly speed up the loading of your submission.</em>
                     <ul>
                         <li>
