+++ DataRepo/urls.py
@@ -5,7 +5,7 @@ from . import views
 urlpatterns = [
     path("", views.home, name="home"),
     path("upload", views.upload, name="upload"),
-    path("validate", views.DataValidationView.as_view(), name="validate"),
+    path("submission", views.DataValidationView.as_view(), name="submission"),
     path("validatedown", views.validation_disabled, name="validatedown"),
     path(
         "search_basic/<str:mdl>/<str:fld>/<str:cmp>/<str:val>/<str:fmt>/",
