+++ DataRepo/views/upload/validation.py
@@ -72,7 +72,7 @@ from DataRepo.utils.text_utils import autowrap
 
 class DataValidationView(FormView):
     form_class = DataSubmissionValidationForm
-    template_name = "DataRepo/validate_submission.html"
+    template_name = "DataRepo/submission.html"
     success_url = ""
     submission_url = settings.SUBMISSION_FORM_URL
     row_key_delim = "__DELIM__"
